Представления для списка всех рецептов (с сортировкой) вынесены в отдельный пакет
